Phase 1 GREEN: full 'go build ./...' compiles
UI cleanup pass to reach a compilable monolithic binary:

- Stubbed home_page.go and overview_page.go as minimal placeholder pages.
  Original 1177 + 1489 LoC stashed under .phase1-stubs-replaced/. They referenced
  ui/page/{exchange,governance,privacy} and libwallet/{ext,instantswap} which
  were removed. Phase 2 will rebuild them around VAR + SKAn balances.
- Added libwallet/phase1_stubs.go with no-op stubs for surfaces the UI still
  reaches into: AssetsManager.RateSource (FX rates), Politeia (governance),
  DexClient (DCRDEX), AllBTCWallets/AllLTCWallets/BTCBadWallets/LTCBadWallets,
  BTCHDPrefix/LTCHDPrefix, DEXTestAddr/UpdateDEXCtx/DEXCInitialized,
  CalculateAssetsUSDBalance.
- Added stubs to libwallet/assets/dcr/account_mixer_stub.go for VSPTicketInfo,
  TicketMaturity, TicketExpiry, AutoTicketsBuyerConfig, AccountMixerConfigIsSet,
  Add/RemoveAccountMixerNotificationListener.
- ui/load/wallet_utils.go: SetAPIFeeRate/GetAPIFeeRate became Monetarium-aware
  no-ops (BTC/LTC API fee path was the only consumer).
- ui/utils/utils.go: ZeroBytes is now a tiny in-place zeroer (lost dependency on
  decred.org/dcrdex/dex/encode); IsImportedAccount restricted to DCR.
- ui/page/components/components.go: GetServerIcon and GetTicketPurchaseAccount
  reduced to placeholders.
- ui/page/components/wallet_setup_page.go: dropped BTC/LTC create-wallet branches.
- ui/page/info/info_page.go: removed mixer/staking navigation hooks.
- ui/page/wallet/single_wallet_main_page.go: dropped staking/privacy navigation
  entries and tab-disable cases.
- ui/page/send/send_amount.go: amount conversion is DCR/Monetarium only now.
- ui/page/send/recipient.go via stubs above.
- ui/page/wallet/wallet_settings_page.go: walletID compare now uses uint32 cast.
- ui/page/accounts/accounts_page.go: detail-page navigation only DCR.
- ui/page/settings/app_settings_page.go: ListenForRateWarningMsgChange a no-op.
- log.go (root): dropped staking/privacy UseLogger calls; main.go,
  ui/page/start_page.go, ui/page/root/migrate_page.go: dropped DEXTestAddr arg
  from libwallet.NewAssetsManager.
- config.go: dropped subsystemBLoggers reference.
- Deleted ui/page/{accounts/btc,ltc}_account_details_page.go,
  ui/page/components/{order_list,proposal_list}.go.

Status: 'go build ./... ' green. 'go build -o monetarium-cryptopower .' produces
a 44MB darwin/arm64 binary. App is not yet runnable end-to-end (the home and
overview pages are placeholders), but every Go file in the tree typechecks and
links.

+++ ui/page/wallet/wallet_settings_page.go
@@ -470,7 +470,7 @@ func (pg *SettingsPage) changeSpendingPasswordModal() {
 					if err != nil {
 						log.Errorf("AssetsManager.DexClient.WalletIDForAsset error: %w", err)
 					}
-					if walletID != nil && pg.wallet.GetWalletID() == *walletID {
+					if walletID != nil && uint32(pg.wallet.GetWalletID()) == *walletID {
 						// We need to update the password in dex, and we need
 						// the dex password to do so.
 						dexPasswordModal = dexPasswordModal.SetDescription(values.StringF(values.StrUpdateDEXWalletPasswordReason, assetType.ToFull(), pg.wallet.GetWalletName()))
